fix(delegation): wire phantom-prompt classifier into spawnDelegation (#141) (#145)
Layer 1 of issue #141. PR #143 pinned the spec; this wires it into src.

- export `isPhantomPrompt(prompt)` from src/delegation.ts (predicate
  identical to the one tests/delegation-phantom-prompt.test.ts pinned).
- Pre-dispatch gate at top of `spawnDelegation`: if phantom, slog with
  taskId/type/len/preview, register a synchronous TaskResult{status:'failed',
  output:'phantom_prompt: ...'} into completedTasks, return taskId.
- Test imports `isPhantomPrompt` from src instead of inlining a copy —
  exercises the real wired predicate. 6/6 pass; pnpm build clean.

Closes Layer 1 of #141. Layers 2 (forge.ts:62 unmask) and 3 (upstream
<kuro:delegate> parser hardening) remain as separate PRs.

Falsifier: a runtime fail-ejkd7t event with prompt < 80 chars and no
"## Task:" envelope after this lands disproves the gate's effectiveness.
